Signal window

+++ gui/windows/algorithm_editor.py
@@ -11,6 +11,7 @@ from PyQt5.QtGui import QFontDatabase, QFontMetricsF, QTextCursor, QKeySequence
 from gui.widgets.QCodeEditor import QCodeEditor
 from gui.widgets.syntax_highlighter import PythonHighlighter
 from gui.src import show_warning_messagebox, show_choose_window
+from gui.themes import code_editor_colors
 from manager.service.global_settings import ALGORITHM_PATH, TICKET_PATH
 from manager.algorithms import Algorithm, check_algorithm_code, execute_algorithm
 from manager.algorithms.algorithm import VALUE_FUNCTIONS, GENERATOR_FUNCTIONS, MULTI_GENERATOR_FUNCTIONS
@@ -84,7 +85,8 @@ class AlgorithmEditor(QDialog):
         
     def setup_code_editor(self, ticket: Union[str, None]) -> None:
         """Set the code editor parameters"""
-        self.code_editor = QCodeEditor()
+        theme_dict = code_editor_colors(self.parent.parent.theme)
+        self.code_editor = QCodeEditor(theme_dict['line_numbers'])
         self.groupBox_code_editor.layout().addWidget(self.code_editor)
         self.set_font_and_highlight(self.code_editor)
         # Tabulation
@@ -102,7 +104,7 @@ class AlgorithmEditor(QDialog):
                 self.code_editor.setPlainText(BASE_ALGORITHM)
         # Displaying algorithm
         self.code_editor.setFocus()
-        self.code_editor.moveCursor(QTextCursor.End)
+        self.code_editor.moveCursor(QTextCursor.Start)
         
     
     def setup_function_lists(self) -> None:
@@ -145,7 +147,10 @@ class AlgorithmEditor(QDialog):
         font = QFontDatabase.systemFont(QFontDatabase.FixedFont)
         plainTextEdit.setFont(font)
         # Highlighting
-        highlight = PythonHighlighter(plainTextEdit.document())
+        highlight = PythonHighlighter(
+            plainTextEdit.document(), 
+            styles=code_editor_colors(self.parent.parent.theme)['syntax_highlight']
+        )
         plainTextEdit.textChanged.connect(lambda: highlight.highlightBlock(None))
         
         
@@ -241,8 +246,15 @@ class AlgorithmEditor(QDialog):
             return
         if self.mode == 'edit':
             self.parent.apply_edit_to_exp_list(alg_ticket)
+        if alg_ticket['name'] in self.parent.parent.protected_algorithms:
+            show_warning_messagebox(self, self.lang_pack.get("alg_protected"))
+            return
+        save_path = os.path.join(ALGORITHM_PATH, alg_ticket['name'] + '.json')
+        if os.path.exists(save_path):
+            if not show_choose_window(self, self.lang_pack.get("an_algorithm") + "'" + alg_ticket['name'] + "'" + self.lang_pack.get('already_exists')):
+                return
         try:  # Saving to file
-            with open(os.path.join(ALGORITHM_PATH, alg_ticket['name'] + '.json'), mode='w', encoding='utf-8') as file:
+            with open(save_path, mode='w', encoding='utf-8') as file:
                 json.dump(alg_ticket, file, ensure_ascii=False, indent=4)
             self.safe_to_close = True
             self.close()
@@ -267,13 +279,4 @@ class AlgorithmEditor(QDialog):
             else:
                 event.accept()
         self.parent.refresh_alg_list()
-        
-        
-# TODO: custom QPlainEditText for tabulation
-        
-# TODO: dark theme?
-# palette = self.code_editor.palette()
-# palette.setColor(QPalette.ColorRole.Base, QColor('#282c34'))
-# palette.setColor(QPalette.ColorRole.Text, QColor('#ffffff'))
-# self.code_editor.setPalette(palette)
